replace google analytics with mixpanel
