+++ packages/core/lib/services/analytics/main.js
@@ -1,5 +1,5 @@
 require("source-map-support/register");
-const analytics = require("./google.js");
+const analytics = require("./mixpanel.js");
 
 const PROCESS_TIMEOUT = 5000; // ms
